HOSPITAL/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pymongo
mysqlclient